Improve the documentation and move the examples
